🧩 BBZQQ1 task: split feedback issue publish helper
